reorganize orphans
